refactor: extract limit helpers into shared/limits.ts
Moves `ALL_MODELS_SENTINEL` and `validateLimitShape` out of the
catch-all `shared/consts.ts` into a dedicated `shared/limits.ts`. The
sentinel name is broad and the validator is limit-specific; they didn't
belong in a grab-bag constants file. Both symbols keep the same
`@shared` import path via the shared barrel.

Also re-runs codegen so the regenerated OpenAPI spec and hey-api client
are in sync with the current route surface.
