+++ platform/shared/hey-api/clients/api/index.ts
@@ -32,7 +32,6 @@ export {
   cerebrasChatCompletionsWithDefaultAgent,
   checkInvitation,
   checkSecretsConnectivity,
-  checkTeamVaultFolderConnectivity,
   cohereChatWithAgent,
   cohereChatWithDefaultAgent,
   completeOnboarding,
@@ -41,14 +40,12 @@ export {
   createChatConversation,
   createChatOpsDmBinding,
   createConnector,
-  createIdentityProvider,
   createInternalMcpCatalogItem,
   createKnowledgeBase,
   createLimit,
   createLlmProviderApiKey,
   createMcpServerInstallationRequest,
   createOptimizationRule,
-  createRole,
   createScheduleTrigger,
   createScheduleTriggerRunConversation,
   createTeam,
@@ -65,7 +62,6 @@ export {
   deleteChatOpsBinding,
   deleteConnector,
   deleteConversationEnabledTools,
-  deleteIdentityProvider,
   deleteIncomingEmailSubscription,
   deleteInternalMcpCatalogItem,
   deleteInternalMcpCatalogItemByName,
@@ -76,10 +72,8 @@ export {
   deleteMcpServerInstallationRequest,
   deleteOptimizationRule,
   deletePendingSignupMember,
-  deleteRole,
   deleteScheduleTrigger,
   deleteTeam,
-  deleteTeamVaultFolder,
   deleteTool,
   deleteToolInvocationPolicy,
   deleteTrustedDataPolicy,
@@ -124,9 +118,6 @@ export {
   getDefaultMcpGateway,
   getDeploymentYamlPreview,
   getHealth,
-  getIdentityProvider,
-  getIdentityProviderIdpLogoutUrl,
-  getIdentityProviders,
   getIncomingEmailStatus,
   getInteraction,
   getInteractionSessions,
@@ -169,7 +160,6 @@ export {
   getOrganizationMembers,
   getOverviewStatistics,
   getPublicConfig,
-  getPublicIdentityProviders,
   getReady,
   getRole,
   getRoles,
@@ -185,8 +175,6 @@ export {
   getTeamMembers,
   getTeamStatistics,
   getTeams,
-  getTeamVaultFolder,
-  getTeamVaultSecretKeys,
   getTokens,
   getTokenValue,
   getToolInvocationPolicies,
@@ -213,7 +201,6 @@ export {
   inspectMcpServer,
   installMcpServer,
   listChatOpsBindings,
-  listTeamVaultFolderSecrets,
   mcpGatewayGet,
   mcpGatewayPost,
   mcpProxyPost,
@@ -255,7 +242,6 @@ export {
   rotateToken,
   rotateUserToken,
   runScheduleTriggerNow,
-  setTeamVaultFolder,
   setupIncomingEmailWebhook,
   shareConversation,
   stopChatStream,
@@ -280,7 +266,6 @@ export {
   updateConnectionSettings,
   updateConnector,
   updateConversationEnabledTools,
-  updateIdentityProvider,
   updateInternalMcpCatalogItem,
   updateKnowledgeBase,
   updateKnowledgeSettings,
@@ -291,7 +276,6 @@ export {
   updateMcpSettings,
   updateModel,
   updateOptimizationRule,
-  updateRole,
   updateScheduleTrigger,
   updateSecuritySettings,
   updateSlackChatOpsConfig,
@@ -456,11 +440,6 @@ export type {
   CheckSecretsConnectivityErrors,
   CheckSecretsConnectivityResponse,
   CheckSecretsConnectivityResponses,
-  CheckTeamVaultFolderConnectivityData,
-  CheckTeamVaultFolderConnectivityError,
-  CheckTeamVaultFolderConnectivityErrors,
-  CheckTeamVaultFolderConnectivityResponse,
-  CheckTeamVaultFolderConnectivityResponses,
   ClientOptions,
   CohereChatRequest,
   CohereChatRequestInput,
@@ -506,11 +485,6 @@ export type {
   CreateConnectorErrors,
   CreateConnectorResponse,
   CreateConnectorResponses,
-  CreateIdentityProviderData,
-  CreateIdentityProviderError,
-  CreateIdentityProviderErrors,
-  CreateIdentityProviderResponse,
-  CreateIdentityProviderResponses,
   CreateInternalMcpCatalogItemData,
   CreateInternalMcpCatalogItemError,
   CreateInternalMcpCatalogItemErrors,
@@ -541,11 +515,6 @@ export type {
   CreateOptimizationRuleErrors,
   CreateOptimizationRuleResponse,
   CreateOptimizationRuleResponses,
-  CreateRoleData,
-  CreateRoleError,
-  CreateRoleErrors,
-  CreateRoleResponse,
-  CreateRoleResponses,
   CreateScheduleTriggerData,
   CreateScheduleTriggerError,
   CreateScheduleTriggerErrors,
@@ -630,11 +599,6 @@ export type {
   DeleteConversationEnabledToolsErrors,
   DeleteConversationEnabledToolsResponse,
   DeleteConversationEnabledToolsResponses,
-  DeleteIdentityProviderData,
-  DeleteIdentityProviderError,
-  DeleteIdentityProviderErrors,
-  DeleteIdentityProviderResponse,
-  DeleteIdentityProviderResponses,
   DeleteIncomingEmailSubscriptionData,
   DeleteIncomingEmailSubscriptionError,
   DeleteIncomingEmailSubscriptionErrors,
@@ -685,11 +649,6 @@ export type {
   DeletePendingSignupMemberErrors,
   DeletePendingSignupMemberResponse,
   DeletePendingSignupMemberResponses,
-  DeleteRoleData,
-  DeleteRoleError,
-  DeleteRoleErrors,
-  DeleteRoleResponse,
-  DeleteRoleResponses,
   DeleteScheduleTriggerData,
   DeleteScheduleTriggerError,
   DeleteScheduleTriggerErrors,
@@ -700,11 +659,6 @@ export type {
   DeleteTeamErrors,
   DeleteTeamResponse,
   DeleteTeamResponses,
-  DeleteTeamVaultFolderData,
-  DeleteTeamVaultFolderError,
-  DeleteTeamVaultFolderErrors,
-  DeleteTeamVaultFolderResponse,
-  DeleteTeamVaultFolderResponses,
   DeleteToolData,
   DeleteToolError,
   DeleteToolErrors,
@@ -919,21 +873,6 @@ export type {
   GetHealthData,
   GetHealthResponse,
   GetHealthResponses,
-  GetIdentityProviderData,
-  GetIdentityProviderError,
-  GetIdentityProviderErrors,
-  GetIdentityProviderIdpLogoutUrlData,
-  GetIdentityProviderIdpLogoutUrlError,
-  GetIdentityProviderIdpLogoutUrlErrors,
-  GetIdentityProviderIdpLogoutUrlResponse,
-  GetIdentityProviderIdpLogoutUrlResponses,
-  GetIdentityProviderResponse,
-  GetIdentityProviderResponses,
-  GetIdentityProvidersData,
-  GetIdentityProvidersError,
-  GetIdentityProvidersErrors,
-  GetIdentityProvidersResponse,
-  GetIdentityProvidersResponses,
   GetIncomingEmailStatusData,
   GetIncomingEmailStatusError,
   GetIncomingEmailStatusErrors,
@@ -1140,11 +1079,6 @@ export type {
   GetPublicConfigData,
   GetPublicConfigResponse,
   GetPublicConfigResponses,
-  GetPublicIdentityProvidersData,
-  GetPublicIdentityProvidersError,
-  GetPublicIdentityProvidersErrors,
-  GetPublicIdentityProvidersResponse,
-  GetPublicIdentityProvidersResponses,
   GetReadyData,
   GetReadyError,
   GetReadyErrors,
@@ -1220,16 +1154,6 @@ export type {
   GetTeamsErrors,
   GetTeamsResponse,
   GetTeamsResponses,
-  GetTeamVaultFolderData,
-  GetTeamVaultFolderError,
-  GetTeamVaultFolderErrors,
-  GetTeamVaultFolderResponse,
-  GetTeamVaultFolderResponses,
-  GetTeamVaultSecretKeysData,
-  GetTeamVaultSecretKeysError,
-  GetTeamVaultSecretKeysErrors,
-  GetTeamVaultSecretKeysResponse,
-  GetTeamVaultSecretKeysResponses,
   GetTokensData,
   GetTokensError,
   GetTokensErrors,
@@ -1358,11 +1282,6 @@ export type {
   ListChatOpsBindingsErrors,
   ListChatOpsBindingsResponse,
   ListChatOpsBindingsResponses,
-  ListTeamVaultFolderSecretsData,
-  ListTeamVaultFolderSecretsError,
-  ListTeamVaultFolderSecretsErrors,
-  ListTeamVaultFolderSecretsResponse,
-  ListTeamVaultFolderSecretsResponses,
   McpGatewayGetData,
   McpGatewayGetError,
   McpGatewayGetErrors,
@@ -1559,11 +1478,6 @@ export type {
   RunScheduleTriggerNowErrors,
   RunScheduleTriggerNowResponse,
   RunScheduleTriggerNowResponses,
-  SetTeamVaultFolderData,
-  SetTeamVaultFolderError,
-  SetTeamVaultFolderErrors,
-  SetTeamVaultFolderResponse,
-  SetTeamVaultFolderResponses,
   SetupIncomingEmailWebhookData,
   SetupIncomingEmailWebhookError,
   SetupIncomingEmailWebhookErrors,
@@ -1682,11 +1596,6 @@ export type {
   UpdateConversationEnabledToolsErrors,
   UpdateConversationEnabledToolsResponse,
   UpdateConversationEnabledToolsResponses,
-  UpdateIdentityProviderData,
-  UpdateIdentityProviderError,
-  UpdateIdentityProviderErrors,
-  UpdateIdentityProviderResponse,
-  UpdateIdentityProviderResponses,
   UpdateInternalMcpCatalogItemData,
   UpdateInternalMcpCatalogItemError,
   UpdateInternalMcpCatalogItemErrors,
@@ -1737,11 +1646,6 @@ export type {
   UpdateOptimizationRuleErrors,
   UpdateOptimizationRuleResponse,
   UpdateOptimizationRuleResponses,
-  UpdateRoleData,
-  UpdateRoleError,
-  UpdateRoleErrors,
-  UpdateRoleResponse,
-  UpdateRoleResponses,
   UpdateScheduleTriggerData,
   UpdateScheduleTriggerError,
   UpdateScheduleTriggerErrors,
