+++ platform/shared/hey-api/clients/api/sdk.gen.ts
@@ -2,7 +2,7 @@
 
 import type { Client, Options as Options2, TDataShape } from './client';
 import { client } from './client.gen';
-import type { AddMcpServerInstallationRequestNoteData, AddMcpServerInstallationRequestNoteErrors, AddMcpServerInstallationRequestNoteResponses, AddTeamExternalGroupData, AddTeamExternalGroupErrors, AddTeamExternalGroupResponses, AddTeamMemberData, AddTeamMemberErrors, AddTeamMemberResponses, AnthropicMessagesWithAgentData, AnthropicMessagesWithAgentErrors, AnthropicMessagesWithAgentResponses, AnthropicMessagesWithDefaultAgentData, AnthropicMessagesWithDefaultAgentErrors, AnthropicMessagesWithDefaultAgentResponses, ApproveMcpServerInstallationRequestData, ApproveMcpServerInstallationRequestErrors, ApproveMcpServerInstallationRequestResponses, AssignConnectorToKnowledgeBasesData, AssignConnectorToKnowledgeBasesErrors, AssignConnectorToKnowledgeBasesResponses, AssignToolToAgentData, AssignToolToAgentErrors, AssignToolToAgentResponses, AutoConfigureAgentToolPoliciesData, AutoConfigureAgentToolPoliciesErrors, AutoConfigureAgentToolPoliciesResponses, AzureChatCompletionsWithAgentData, AzureChatCompletionsWithAgentErrors, AzureChatCompletionsWithAgentResponses, AzureChatCompletionsWithDefaultAgentData, AzureChatCompletionsWithDefaultAgentErrors, AzureChatCompletionsWithDefaultAgentResponses, AzureResponsesWithAgentData, AzureResponsesWithAgentErrors, AzureResponsesWithAgentResponses, AzureResponsesWithDefaultAgentData, AzureResponsesWithDefaultAgentErrors, AzureResponsesWithDefaultAgentResponses, BedrockConverseStreamWithAgentAndModelData, BedrockConverseStreamWithAgentAndModelResponses, BedrockConverseStreamWithAgentData, BedrockConverseStreamWithAgentResponses, BedrockConverseStreamWithDefaultAgentData, BedrockConverseStreamWithDefaultAgentResponses, BedrockConverseWithAgentAndModelData, BedrockConverseWithAgentAndModelErrors, BedrockConverseWithAgentAndModelResponses, BedrockConverseWithAgentData, BedrockConverseWithAgentErrors, BedrockConverseWithAgentResponses, BedrockConverseWithDefaultAgentData, BedrockConverseWithDefaultAgentErrors, BedrockConverseWithDefaultAgentResponses, BedrockOpenaiChatCompletionsWithAgentData, BedrockOpenaiChatCompletionsWithAgentResponses, BedrockOpenaiChatCompletionsWithDefaultAgentData, BedrockOpenaiChatCompletionsWithDefaultAgentResponses, BedrockOpenaiListModelsWithAgentData, BedrockOpenaiListModelsWithAgentErrors, BedrockOpenaiListModelsWithAgentResponses, BedrockOpenaiListModelsWithDefaultAgentData, BedrockOpenaiListModelsWithDefaultAgentErrors, BedrockOpenaiListModelsWithDefaultAgentResponses, BulkAssignToolsData, BulkAssignToolsErrors, BulkAssignToolsResponses, BulkUpdateChatOpsBindingsData, BulkUpdateChatOpsBindingsErrors, BulkUpdateChatOpsBindingsResponses, BulkUpsertDefaultCallPolicyData, BulkUpsertDefaultCallPolicyErrors, BulkUpsertDefaultCallPolicyResponses, BulkUpsertDefaultResultPolicyData, BulkUpsertDefaultResultPolicyErrors, BulkUpsertDefaultResultPolicyResponses, CerebrasChatCompletionsWithAgentData, CerebrasChatCompletionsWithAgentErrors, CerebrasChatCompletionsWithAgentResponses, CerebrasChatCompletionsWithDefaultAgentData, CerebrasChatCompletionsWithDefaultAgentErrors, CerebrasChatCompletionsWithDefaultAgentResponses, CheckInvitationData, CheckInvitationErrors, CheckInvitationResponses, CheckSecretsConnectivityData, CheckSecretsConnectivityErrors, CheckSecretsConnectivityResponses, CheckTeamVaultFolderConnectivityData, CheckTeamVaultFolderConnectivityErrors, CheckTeamVaultFolderConnectivityResponses, CohereChatWithAgentData, CohereChatWithAgentErrors, CohereChatWithAgentResponses, CohereChatWithDefaultAgentData, CohereChatWithDefaultAgentErrors, CohereChatWithDefaultAgentResponses, CompleteOnboardingData, CompleteOnboardingErrors, CompleteOnboardingResponses, CreateAgentData, CreateAgentErrors, CreateAgentResponses, CreateApiKeyData, CreateApiKeyErrors, CreateApiKeyResponses, CreateChatConversationData, CreateChatConversationErrors, CreateChatConversationResponses, CreateChatOpsDmBindingData, CreateChatOpsDmBindingErrors, CreateChatOpsDmBindingResponses, CreateConnectorData, CreateConnectorErrors, CreateConnectorResponses, CreateIdentityProviderData, CreateIdentityProviderErrors, CreateIdentityProviderResponses, CreateInternalMcpCatalogItemData, CreateInternalMcpCatalogItemErrors, CreateInternalMcpCatalogItemResponses, CreateKnowledgeBaseData, CreateKnowledgeBaseErrors, CreateKnowledgeBaseResponses, CreateLimitData, CreateLimitErrors, CreateLimitResponses, CreateLlmProviderApiKeyData, CreateLlmProviderApiKeyErrors, CreateLlmProviderApiKeyResponses, CreateMcpServerInstallationRequestData, CreateMcpServerInstallationRequestErrors, CreateMcpServerInstallationRequestResponses, CreateOptimizationRuleData, CreateOptimizationRuleErrors, CreateOptimizationRuleResponses, CreateRoleData, CreateRoleErrors, CreateRoleResponses, CreateScheduleTriggerData, CreateScheduleTriggerErrors, CreateScheduleTriggerResponses, CreateScheduleTriggerRunConversationData, CreateScheduleTriggerRunConversationErrors, CreateScheduleTriggerRunConversationResponses, CreateTeamData, CreateTeamErrors, CreateTeamResponses, CreateToolInvocationPolicyData, CreateToolInvocationPolicyErrors, CreateToolInvocationPolicyResponses, CreateTrustedDataPolicyData, CreateTrustedDataPolicyErrors, CreateTrustedDataPolicyResponses, CreateVirtualApiKeyData, CreateVirtualApiKeyErrors, CreateVirtualApiKeyResponses, DeclineMcpServerInstallationRequestData, DeclineMcpServerInstallationRequestErrors, DeclineMcpServerInstallationRequestResponses, DeepseekChatCompletionsWithAgentData, DeepseekChatCompletionsWithAgentErrors, DeepseekChatCompletionsWithAgentResponses, DeepseekChatCompletionsWithDefaultAgentData, DeepseekChatCompletionsWithDefaultAgentErrors, DeepseekChatCompletionsWithDefaultAgentResponses, DeleteAgentData, DeleteAgentDelegationData, DeleteAgentDelegationErrors, DeleteAgentDelegationResponses, DeleteAgentErrors, DeleteAgentResponses, DeleteApiKeyData, DeleteApiKeyErrors, DeleteApiKeyResponses, DeleteChatConversationData, DeleteChatConversationErrors, DeleteChatConversationResponses, DeleteChatOpsBindingData, DeleteChatOpsBindingErrors, DeleteChatOpsBindingResponses, DeleteConnectorData, DeleteConnectorErrors, DeleteConnectorResponses, DeleteConversationEnabledToolsData, DeleteConversationEnabledToolsErrors, DeleteConversationEnabledToolsResponses, DeleteIdentityProviderData, DeleteIdentityProviderErrors, DeleteIdentityProviderResponses, DeleteIncomingEmailSubscriptionData, DeleteIncomingEmailSubscriptionErrors, DeleteIncomingEmailSubscriptionResponses, DeleteInternalMcpCatalogItemByNameData, DeleteInternalMcpCatalogItemByNameErrors, DeleteInternalMcpCatalogItemByNameResponses, DeleteInternalMcpCatalogItemData, DeleteInternalMcpCatalogItemErrors, DeleteInternalMcpCatalogItemResponses, DeleteKnowledgeBaseData, DeleteKnowledgeBaseErrors, DeleteKnowledgeBaseResponses, DeleteLimitData, DeleteLimitErrors, DeleteLimitResponses, DeleteLlmProviderApiKeyData, DeleteLlmProviderApiKeyErrors, DeleteLlmProviderApiKeyResponses, DeleteMcpServerData, DeleteMcpServerErrors, DeleteMcpServerInstallationRequestData, DeleteMcpServerInstallationRequestErrors, DeleteMcpServerInstallationRequestResponses, DeleteMcpServerResponses, DeleteOptimizationRuleData, DeleteOptimizationRuleErrors, DeleteOptimizationRuleResponses, DeletePendingSignupMemberData, DeletePendingSignupMemberErrors, DeletePendingSignupMemberResponses, DeleteRoleData, DeleteRoleErrors, DeleteRoleResponses, DeleteScheduleTriggerData, DeleteScheduleTriggerErrors, DeleteScheduleTriggerResponses, DeleteTeamData, DeleteTeamErrors, DeleteTeamResponses, DeleteTeamVaultFolderData, DeleteTeamVaultFolderErrors, DeleteTeamVaultFolderResponses, DeleteToolData, DeleteToolErrors, DeleteToolInvocationPolicyData, DeleteToolInvocationPolicyErrors, DeleteToolInvocationPolicyResponses, DeleteToolResponses, DeleteTrustedDataPolicyData, DeleteTrustedDataPolicyErrors, DeleteTrustedDataPolicyResponses, DeleteVirtualApiKeyData, DeleteVirtualApiKeyErrors, DeleteVirtualApiKeyResponses, DisableScheduleTriggerData, DisableScheduleTriggerErrors, DisableScheduleTriggerResponses, DropEmbeddingConfigData, DropEmbeddingConfigErrors, DropEmbeddingConfigResponses, EnableScheduleTriggerData, EnableScheduleTriggerErrors, EnableScheduleTriggerResponses, ForceResyncConnectorData, ForceResyncConnectorErrors, ForceResyncConnectorResponses, ForkSharedConversationData, ForkSharedConversationErrors, ForkSharedConversationResponses, GenerateChatConversationTitleData, GenerateChatConversationTitleErrors, GenerateChatConversationTitleResponses, GetAgentData, GetAgentDelegationsData, GetAgentDelegationsErrors, GetAgentDelegationsResponses, GetAgentEmailAddressData, GetAgentEmailAddressErrors, GetAgentEmailAddressResponses, GetAgentErrors, GetAgentResponses, GetAgentsData, GetAgentsErrors, GetAgentsResponses, GetAgentStatisticsData, GetAgentStatisticsErrors, GetAgentStatisticsResponses, GetAgentToolsData, GetAgentToolsErrors, GetAgentToolsResponses, GetAllAgentsData, GetAllAgentsErrors, GetAllAgentsResponses, GetAllAgentToolsData, GetAllAgentToolsErrors, GetAllAgentToolsResponses, GetAllDelegationConnectionsData, GetAllDelegationConnectionsErrors, GetAllDelegationConnectionsResponses, GetAllVirtualApiKeysData, GetAllVirtualApiKeysErrors, GetAllVirtualApiKeysResponses, GetApiAuthBy__Data, GetApiAuthBy__Responses, GetApiAuthOauth2AuthorizeData, GetApiAuthOauth2AuthorizeResponses, GetApiKeyData, GetApiKeyErrors, GetApiKeyResponses, GetApiKeysData, GetApiKeysErrors, GetApiKeysResponses, GetAppearanceSettingsData, GetAppearanceSettingsErrors, GetAppearanceSettingsResponses, GetAvailableLlmProviderApiKeysData, GetAvailableLlmProviderApiKeysErrors, GetAvailableLlmProviderApiKeysResponses, GetChatAgentMcpToolsData, GetChatAgentMcpToolsErrors, GetChatAgentMcpToolsResponses, GetChatConversationData, GetChatConversationErrors, GetChatConversationResponses, GetChatConversationsData, GetChatConversationsErrors, GetChatConversationsResponses, GetChatOpsStatusData, GetChatOpsStatusErrors, GetChatOpsStatusResponses, GetConfigData, GetConfigResponses, GetConnectorData, GetConnectorErrors, GetConnectorKnowledgeBasesData, GetConnectorKnowledgeBasesErrors, GetConnectorKnowledgeBasesResponses, GetConnectorResponses, GetConnectorRunData, GetConnectorRunErrors, GetConnectorRunResponses, GetConnectorRunsData, GetConnectorRunsErrors, GetConnectorRunsResponses, GetConnectorsData, GetConnectorsErrors, GetConnectorsResponses, GetConversationEnabledToolsData, GetConversationEnabledToolsErrors, GetConversationEnabledToolsResponses, GetConversationShareData, GetConversationShareErrors, GetConversationShareResponses, GetCostSavingsStatisticsData, GetCostSavingsStatisticsErrors, GetCostSavingsStatisticsResponses, GetDefaultCredentialsStatusData, GetDefaultCredentialsStatusErrors, GetDefaultCredentialsStatusResponses, GetDefaultLlmProxyData, GetDefaultLlmProxyErrors, GetDefaultLlmProxyResponses, GetDefaultMcpGatewayData, GetDefaultMcpGatewayErrors, GetDefaultMcpGatewayResponses, GetDeploymentYamlPreviewData, GetDeploymentYamlPreviewErrors, GetDeploymentYamlPreviewResponses, GetHealthData, GetHealthResponses, GetIdentityProviderData, GetIdentityProviderErrors, GetIdentityProviderIdpLogoutUrlData, GetIdentityProviderIdpLogoutUrlErrors, GetIdentityProviderIdpLogoutUrlResponses, GetIdentityProviderResponses, GetIdentityProvidersData, GetIdentityProvidersErrors, GetIdentityProvidersResponses, GetIncomingEmailStatusData, GetIncomingEmailStatusErrors, GetIncomingEmailStatusResponses, GetInteractionData, GetInteractionErrors, GetInteractionResponses, GetInteractionsData, GetInteractionsErrors, GetInteractionSessionsData, GetInteractionSessionsErrors, GetInteractionSessionsResponses, GetInteractionsResponses, GetInternalMcpCatalogData, GetInternalMcpCatalogErrors, GetInternalMcpCatalogItemData, GetInternalMcpCatalogItemErrors, GetInternalMcpCatalogItemResponses, GetInternalMcpCatalogLabelKeysData, GetInternalMcpCatalogLabelKeysErrors, GetInternalMcpCatalogLabelKeysResponses, GetInternalMcpCatalogLabelValuesData, GetInternalMcpCatalogLabelValuesErrors, GetInternalMcpCatalogLabelValuesResponses, GetInternalMcpCatalogResponses, GetInternalMcpCatalogToolsData, GetInternalMcpCatalogToolsErrors, GetInternalMcpCatalogToolsResponses, GetK8sImagePullSecretsData, GetK8sImagePullSecretsErrors, GetK8sImagePullSecretsResponses, GetKnowledgeBaseData, GetKnowledgeBaseErrors, GetKnowledgeBaseHealthData, GetKnowledgeBaseHealthErrors, GetKnowledgeBaseHealthResponses, GetKnowledgeBaseResponses, GetKnowledgeBasesData, GetKnowledgeBasesErrors, GetKnowledgeBasesResponses, GetLabelKeysData, GetLabelKeysErrors, GetLabelKeysResponses, GetLabelValuesData, GetLabelValuesErrors, GetLabelValuesResponses, GetLimitData, GetLimitErrors, GetLimitResponses, GetLimitsData, GetLimitsErrors, GetLimitsResponses, GetLlmModelsData, GetLlmModelsErrors, GetLlmModelsResponses, GetLlmProviderApiKeyData, GetLlmProviderApiKeyErrors, GetLlmProviderApiKeyResponses, GetLlmProviderApiKeysData, GetLlmProviderApiKeysErrors, GetLlmProviderApiKeysResponses, GetMcpServerData, GetMcpServerErrors, GetMcpServerInstallationRequestData, GetMcpServerInstallationRequestErrors, GetMcpServerInstallationRequestResponses, GetMcpServerInstallationRequestsData, GetMcpServerInstallationRequestsErrors, GetMcpServerInstallationRequestsResponses, GetMcpServerInstallationStatusData, GetMcpServerInstallationStatusErrors, GetMcpServerInstallationStatusResponses, GetMcpServerResponses, GetMcpServersData, GetMcpServersErrors, GetMcpServersResponses, GetMcpServerToolsData, GetMcpServerToolsErrors, GetMcpServerToolsResponses, GetMcpToolCallData, GetMcpToolCallErrors, GetMcpToolCallResponses, GetMcpToolCallsData, GetMcpToolCallsErrors, GetMcpToolCallsResponses, GetMemberDefaultAgentData, GetMemberDefaultAgentErrors, GetMemberDefaultAgentResponses, GetMembersData, GetMembersErrors, GetMemberSignupStatusData, GetMemberSignupStatusErrors, GetMemberSignupStatusResponses, GetMembersResponses, GetModelStatisticsData, GetModelStatisticsErrors, GetModelStatisticsResponses, GetModelsWithApiKeysData, GetModelsWithApiKeysErrors, GetModelsWithApiKeysResponses, GetOAuthClientInfoData, GetOAuthClientInfoResponses, GetOnboardingStatusData, GetOnboardingStatusErrors, GetOnboardingStatusResponses, GetOperatorsData, GetOperatorsErrors, GetOperatorsResponses, GetOptimizationRulesData, GetOptimizationRulesErrors, GetOptimizationRulesResponses, GetOrganizationData, GetOrganizationErrors, GetOrganizationMemberData, GetOrganizationMemberErrors, GetOrganizationMemberResponses, GetOrganizationMembersData, GetOrganizationMembersErrors, GetOrganizationMembersResponses, GetOrganizationResponses, GetOverviewStatisticsData, GetOverviewStatisticsErrors, GetOverviewStatisticsResponses, GetPublicConfigData, GetPublicConfigResponses, GetPublicIdentityProvidersData, GetPublicIdentityProvidersErrors, GetPublicIdentityProvidersResponses, GetReadyData, GetReadyErrors, GetReadyResponses, GetRoleData, GetRoleErrors, GetRoleResponses, GetRolesData, GetRolesErrors, GetRolesResponses, GetScheduleTriggerData, GetScheduleTriggerErrors, GetScheduleTriggerResponses, GetScheduleTriggerRunData, GetScheduleTriggerRunErrors, GetScheduleTriggerRunResponses, GetScheduleTriggerRunsData, GetScheduleTriggerRunsErrors, GetScheduleTriggerRunsResponses, GetScheduleTriggersData, GetScheduleTriggersErrors, GetScheduleTriggersResponses, GetSecretData, GetSecretErrors, GetSecretResponses, GetSecretsTypeData, GetSecretsTypeErrors, GetSecretsTypeResponses, GetSharedConversationData, GetSharedConversationErrors, GetSharedConversationResponses, GetTeamData, GetTeamErrors, GetTeamExternalGroupsData, GetTeamExternalGroupsErrors, GetTeamExternalGroupsResponses, GetTeamMembersData, GetTeamMembersErrors, GetTeamMembersResponses, GetTeamResponses, GetTeamsData, GetTeamsErrors, GetTeamsResponses, GetTeamStatisticsData, GetTeamStatisticsErrors, GetTeamStatisticsResponses, GetTeamVaultFolderData, GetTeamVaultFolderErrors, GetTeamVaultFolderResponses, GetTeamVaultSecretKeysData, GetTeamVaultSecretKeysErrors, GetTeamVaultSecretKeysResponses, GetTokensData, GetTokensErrors, GetTokensResponses, GetTokenValueData, GetTokenValueErrors, GetTokenValueResponses, GetToolInvocationPoliciesData, GetToolInvocationPoliciesErrors, GetToolInvocationPoliciesResponses, GetToolInvocationPolicyData, GetToolInvocationPolicyErrors, GetToolInvocationPolicyResponses, GetToolsData, GetToolsErrors, GetToolsResponses, GetToolsWithAssignmentsData, GetToolsWithAssignmentsErrors, GetToolsWithAssignmentsResponses, GetTrustedDataPoliciesData, GetTrustedDataPoliciesErrors, GetTrustedDataPoliciesResponses, GetTrustedDataPolicyData, GetTrustedDataPolicyErrors, GetTrustedDataPolicyResponses, GetUniqueExternalAgentIdsData, GetUniqueExternalAgentIdsErrors, GetUniqueExternalAgentIdsResponses, GetUniqueUserIdsData, GetUniqueUserIdsErrors, GetUniqueUserIdsResponses, GetUserPermissionsData, GetUserPermissionsErrors, GetUserPermissionsResponses, GetUserStatisticsData, GetUserStatisticsErrors, GetUserStatisticsResponses, GetUserTokenData, GetUserTokenErrors, GetUserTokenResponses, GetUserTokenValueData, GetUserTokenValueErrors, GetUserTokenValueResponses, GetV1A2aByAgentIdWellKnownAgentJsonData, GetV1A2aByAgentIdWellKnownAgentJsonResponses, GetVirtualApiKeysData, GetVirtualApiKeysErrors, GetVirtualApiKeysResponses, GetVirtualKeyStatisticsData, GetVirtualKeyStatisticsErrors, GetVirtualKeyStatisticsResponses, GetWellKnownOauthAuthorizationServerData, GetWellKnownOauthAuthorizationServerResponses, GetWellKnownOauthProtectedResourceBy__Data, GetWellKnownOauthProtectedResourceBy__Responses, GroqChatCompletionsWithAgentData, GroqChatCompletionsWithAgentErrors, GroqChatCompletionsWithAgentResponses, GroqChatCompletionsWithDefaultAgentData, GroqChatCompletionsWithDefaultAgentErrors, GroqChatCompletionsWithDefaultAgentResponses, HandleOAuthCallbackData, HandleOAuthCallbackErrors, HandleOAuthCallbackResponses, InitiateOAuthData, InitiateOAuthErrors, InitiateOAuthResponses, InspectMcpServerData, InspectMcpServerErrors, InspectMcpServerResponses, InstallMcpServerData, InstallMcpServerErrors, InstallMcpServerResponses, ListChatOpsBindingsData, ListChatOpsBindingsErrors, ListChatOpsBindingsResponses, ListTeamVaultFolderSecretsData, ListTeamVaultFolderSecretsErrors, ListTeamVaultFolderSecretsResponses, McpGatewayGetData, McpGatewayGetErrors, McpGatewayGetResponses, McpGatewayPostData, McpGatewayPostResponses, McpProxyPostData, McpProxyPostResponses, MinimaxChatCompletionsWithAgentData, MinimaxChatCompletionsWithAgentErrors, MinimaxChatCompletionsWithAgentResponses, MinimaxChatCompletionsWithDefaultAgentData, MinimaxChatCompletionsWithDefaultAgentErrors, MinimaxChatCompletionsWithDefaultAgentResponses, MistralChatCompletionsWithAgentData, MistralChatCompletionsWithAgentErrors, MistralChatCompletionsWithAgentResponses, MistralChatCompletionsWithDefaultAgentData, MistralChatCompletionsWithDefaultAgentErrors, MistralChatCompletionsWithDefaultAgentResponses, OllamaChatCompletionsWithAgentData, OllamaChatCompletionsWithAgentErrors, OllamaChatCompletionsWithAgentResponses, OllamaChatCompletionsWithDefaultAgentData, OllamaChatCompletionsWithDefaultAgentErrors, OllamaChatCompletionsWithDefaultAgentResponses, OpenAiChatCompletionsWithAgentData, OpenAiChatCompletionsWithAgentErrors, OpenAiChatCompletionsWithAgentResponses, OpenAiChatCompletionsWithDefaultAgentData, OpenAiChatCompletionsWithDefaultAgentErrors, OpenAiChatCompletionsWithDefaultAgentResponses, OpenrouterChatCompletionsWithAgentData, OpenrouterChatCompletionsWithAgentErrors, OpenrouterChatCompletionsWithAgentResponses, OpenrouterChatCompletionsWithDefaultAgentData, OpenrouterChatCompletionsWithDefaultAgentErrors, OpenrouterChatCompletionsWithDefaultAgentResponses, PerplexityChatCompletionsWithAgentData, PerplexityChatCompletionsWithAgentErrors, PerplexityChatCompletionsWithAgentResponses, PerplexityChatCompletionsWithDefaultAgentData, PerplexityChatCompletionsWithDefaultAgentErrors, PerplexityChatCompletionsWithDefaultAgentResponses, PostApiAuthBy__Data, PostApiAuthBy__Responses, PostApiAuthOauth2RegisterData, PostApiAuthOauth2RegisterResponses, PostApiAuthOauth2TokenData, PostApiAuthOauth2TokenResponses, PostApiAuthOrganizationRemoveMemberData, PostApiAuthOrganizationRemoveMemberResponses, PostApiAuthSignInSsoData, PostApiAuthSignInSsoResponses, PostApiWebhooksChatopsMsTeamsData, PostApiWebhooksChatopsMsTeamsErrors, PostApiWebhooksChatopsMsTeamsResponses, PostApiWebhooksChatopsSlackData, PostApiWebhooksChatopsSlackErrors, PostApiWebhooksChatopsSlackInteractiveData, PostApiWebhooksChatopsSlackInteractiveErrors, PostApiWebhooksChatopsSlackInteractiveResponses, PostApiWebhooksChatopsSlackResponses, PostApiWebhooksChatopsSlackSlashCommandData, PostApiWebhooksChatopsSlackSlashCommandErrors, PostApiWebhooksChatopsSlackSlashCommandResponses, PostApiWebhooksIncomingEmailData, PostApiWebhooksIncomingEmailErrors, PostApiWebhooksIncomingEmailResponses, PostV1A2aByAgentIdData, PostV1A2aByAgentIdResponses, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentData, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentErrors, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentResponses, PostV1GeminiByAgentIdV1BetaModelsByModelStreamGenerateContentData, PostV1GeminiByAgentIdV1BetaModelsByModelStreamGenerateContentErrors, PostV1GeminiV1BetaModelsByModelGenerateContentData, PostV1GeminiV1BetaModelsByModelGenerateContentErrors, PostV1GeminiV1BetaModelsByModelGenerateContentResponses, PostV1GeminiV1BetaModelsByModelStreamGenerateContentData, PostV1GeminiV1BetaModelsByModelStreamGenerateContentErrors, ReauthenticateMcpServerData, ReauthenticateMcpServerErrors, ReauthenticateMcpServerResponses, RefreshChatOpsChannelDiscoveryData, RefreshChatOpsChannelDiscoveryErrors, RefreshChatOpsChannelDiscoveryResponses, ReinstallMcpServerData, ReinstallMcpServerErrors, ReinstallMcpServerResponses, RemoveTeamExternalGroupData, RemoveTeamExternalGroupErrors, RemoveTeamExternalGroupResponses, RemoveTeamMemberData, RemoveTeamMemberErrors, RemoveTeamMemberResponses, RenewIncomingEmailSubscriptionData, RenewIncomingEmailSubscriptionErrors, RenewIncomingEmailSubscriptionResponses, ResetDeploymentYamlData, ResetDeploymentYamlErrors, ResetDeploymentYamlResponses, RotateTokenData, RotateTokenErrors, RotateTokenResponses, RotateUserTokenData, RotateUserTokenErrors, RotateUserTokenResponses, RunScheduleTriggerNowData, RunScheduleTriggerNowErrors, RunScheduleTriggerNowResponses, SetTeamVaultFolderData, SetTeamVaultFolderErrors, SetTeamVaultFolderResponses, SetupIncomingEmailWebhookData, SetupIncomingEmailWebhookErrors, SetupIncomingEmailWebhookResponses, ShareConversationData, ShareConversationErrors, ShareConversationResponses, StopChatStreamData, StopChatStreamErrors, StopChatStreamResponses, StreamChatData, StreamChatErrors, SubmitOAuthConsentData, SubmitOAuthConsentErrors, SubmitOAuthConsentResponses, SyncAgentDelegationsData, SyncAgentDelegationsErrors, SyncAgentDelegationsResponses, SyncConnectorData, SyncConnectorErrors, SyncConnectorResponses, SyncLlmModelsData, SyncLlmModelsErrors, SyncLlmModelsResponses, TestConnectorConnectionData, TestConnectorConnectionErrors, TestConnectorConnectionResponses, TestEmbeddingConnectionData, TestEmbeddingConnectionErrors, TestEmbeddingConnectionResponses, UnassignConnectorFromKnowledgeBaseData, UnassignConnectorFromKnowledgeBaseErrors, UnassignConnectorFromKnowledgeBaseResponses, UnassignToolFromAgentData, UnassignToolFromAgentErrors, UnassignToolFromAgentResponses, UnshareConversationData, UnshareConversationErrors, UnshareConversationResponses, UpdateAgentData, UpdateAgentErrors, UpdateAgentResponses, UpdateAgentSettingsData, UpdateAgentSettingsErrors, UpdateAgentSettingsResponses, UpdateAgentToolData, UpdateAgentToolErrors, UpdateAgentToolResponses, UpdateAppearanceSettingsData, UpdateAppearanceSettingsErrors, UpdateAppearanceSettingsResponses, UpdateChatConversationData, UpdateChatConversationErrors, UpdateChatConversationResponses, UpdateChatMessageData, UpdateChatMessageErrors, UpdateChatMessageResponses, UpdateChatOpsBindingData, UpdateChatOpsBindingErrors, UpdateChatOpsBindingResponses, UpdateChatOpsConfigInQuickstartData, UpdateChatOpsConfigInQuickstartErrors, UpdateChatOpsConfigInQuickstartResponses, UpdateConnectionSettingsData, UpdateConnectionSettingsErrors, UpdateConnectionSettingsResponses, UpdateConnectorData, UpdateConnectorErrors, UpdateConnectorResponses, UpdateConversationEnabledToolsData, UpdateConversationEnabledToolsErrors, UpdateConversationEnabledToolsResponses, UpdateIdentityProviderData, UpdateIdentityProviderErrors, UpdateIdentityProviderResponses, UpdateInternalMcpCatalogItemData, UpdateInternalMcpCatalogItemErrors, UpdateInternalMcpCatalogItemResponses, UpdateKnowledgeBaseData, UpdateKnowledgeBaseErrors, UpdateKnowledgeBaseResponses, UpdateKnowledgeSettingsData, UpdateKnowledgeSettingsErrors, UpdateKnowledgeSettingsResponses, UpdateLimitData, UpdateLimitErrors, UpdateLimitResponses, UpdateLlmProviderApiKeyData, UpdateLlmProviderApiKeyErrors, UpdateLlmProviderApiKeyResponses, UpdateLlmSettingsData, UpdateLlmSettingsErrors, UpdateLlmSettingsResponses, UpdateMcpServerInstallationRequestData, UpdateMcpServerInstallationRequestErrors, UpdateMcpServerInstallationRequestResponses, UpdateMcpSettingsData, UpdateMcpSettingsErrors, UpdateMcpSettingsResponses, UpdateModelData, UpdateModelErrors, UpdateModelResponses, UpdateOptimizationRuleData, UpdateOptimizationRuleErrors, UpdateOptimizationRuleResponses, UpdateRoleData, UpdateRoleErrors, UpdateRoleResponses, UpdateScheduleTriggerData, UpdateScheduleTriggerErrors, UpdateScheduleTriggerResponses, UpdateSecuritySettingsData, UpdateSecuritySettingsErrors, UpdateSecuritySettingsResponses, UpdateSlackChatOpsConfigData, UpdateSlackChatOpsConfigErrors, UpdateSlackChatOpsConfigResponses, UpdateTeamData, UpdateTeamErrors, UpdateTeamResponses, UpdateToolInvocationPolicyData, UpdateToolInvocationPolicyErrors, UpdateToolInvocationPolicyResponses, UpdateTrustedDataPolicyData, UpdateTrustedDataPolicyErrors, UpdateTrustedDataPolicyResponses, UpdateVirtualApiKeyData, UpdateVirtualApiKeyErrors, UpdateVirtualApiKeyResponses, ValidateDeploymentYamlData, ValidateDeploymentYamlErrors, ValidateDeploymentYamlResponses, VllmChatCompletionsWithAgentData, VllmChatCompletionsWithAgentErrors, VllmChatCompletionsWithAgentResponses, VllmChatCompletionsWithDefaultAgentData, VllmChatCompletionsWithDefaultAgentErrors, VllmChatCompletionsWithDefaultAgentResponses, XaiChatCompletionsWithAgentData, XaiChatCompletionsWithAgentErrors, XaiChatCompletionsWithAgentResponses, XaiChatCompletionsWithDefaultAgentData, XaiChatCompletionsWithDefaultAgentErrors, XaiChatCompletionsWithDefaultAgentResponses, ZhipuaiChatCompletionsWithAgentData, ZhipuaiChatCompletionsWithAgentErrors, ZhipuaiChatCompletionsWithAgentResponses, ZhipuaiChatCompletionsWithDefaultAgentData, ZhipuaiChatCompletionsWithDefaultAgentErrors, ZhipuaiChatCompletionsWithDefaultAgentResponses } from './types.gen';
+import type { AddMcpServerInstallationRequestNoteData, AddMcpServerInstallationRequestNoteErrors, AddMcpServerInstallationRequestNoteResponses, AddTeamExternalGroupData, AddTeamExternalGroupErrors, AddTeamExternalGroupResponses, AddTeamMemberData, AddTeamMemberErrors, AddTeamMemberResponses, AnthropicMessagesWithAgentData, AnthropicMessagesWithAgentErrors, AnthropicMessagesWithAgentResponses, AnthropicMessagesWithDefaultAgentData, AnthropicMessagesWithDefaultAgentErrors, AnthropicMessagesWithDefaultAgentResponses, ApproveMcpServerInstallationRequestData, ApproveMcpServerInstallationRequestErrors, ApproveMcpServerInstallationRequestResponses, AssignConnectorToKnowledgeBasesData, AssignConnectorToKnowledgeBasesErrors, AssignConnectorToKnowledgeBasesResponses, AssignToolToAgentData, AssignToolToAgentErrors, AssignToolToAgentResponses, AutoConfigureAgentToolPoliciesData, AutoConfigureAgentToolPoliciesErrors, AutoConfigureAgentToolPoliciesResponses, AzureChatCompletionsWithAgentData, AzureChatCompletionsWithAgentErrors, AzureChatCompletionsWithAgentResponses, AzureChatCompletionsWithDefaultAgentData, AzureChatCompletionsWithDefaultAgentErrors, AzureChatCompletionsWithDefaultAgentResponses, AzureResponsesWithAgentData, AzureResponsesWithAgentErrors, AzureResponsesWithAgentResponses, AzureResponsesWithDefaultAgentData, AzureResponsesWithDefaultAgentErrors, AzureResponsesWithDefaultAgentResponses, BedrockConverseStreamWithAgentAndModelData, BedrockConverseStreamWithAgentAndModelResponses, BedrockConverseStreamWithAgentData, BedrockConverseStreamWithAgentResponses, BedrockConverseStreamWithDefaultAgentData, BedrockConverseStreamWithDefaultAgentResponses, BedrockConverseWithAgentAndModelData, BedrockConverseWithAgentAndModelErrors, BedrockConverseWithAgentAndModelResponses, BedrockConverseWithAgentData, BedrockConverseWithAgentErrors, BedrockConverseWithAgentResponses, BedrockConverseWithDefaultAgentData, BedrockConverseWithDefaultAgentErrors, BedrockConverseWithDefaultAgentResponses, BedrockOpenaiChatCompletionsWithAgentData, BedrockOpenaiChatCompletionsWithAgentResponses, BedrockOpenaiChatCompletionsWithDefaultAgentData, BedrockOpenaiChatCompletionsWithDefaultAgentResponses, BedrockOpenaiListModelsWithAgentData, BedrockOpenaiListModelsWithAgentErrors, BedrockOpenaiListModelsWithAgentResponses, BedrockOpenaiListModelsWithDefaultAgentData, BedrockOpenaiListModelsWithDefaultAgentErrors, BedrockOpenaiListModelsWithDefaultAgentResponses, BulkAssignToolsData, BulkAssignToolsErrors, BulkAssignToolsResponses, BulkUpdateChatOpsBindingsData, BulkUpdateChatOpsBindingsErrors, BulkUpdateChatOpsBindingsResponses, BulkUpsertDefaultCallPolicyData, BulkUpsertDefaultCallPolicyErrors, BulkUpsertDefaultCallPolicyResponses, BulkUpsertDefaultResultPolicyData, BulkUpsertDefaultResultPolicyErrors, BulkUpsertDefaultResultPolicyResponses, CerebrasChatCompletionsWithAgentData, CerebrasChatCompletionsWithAgentErrors, CerebrasChatCompletionsWithAgentResponses, CerebrasChatCompletionsWithDefaultAgentData, CerebrasChatCompletionsWithDefaultAgentErrors, CerebrasChatCompletionsWithDefaultAgentResponses, CheckInvitationData, CheckInvitationErrors, CheckInvitationResponses, CheckSecretsConnectivityData, CheckSecretsConnectivityErrors, CheckSecretsConnectivityResponses, CohereChatWithAgentData, CohereChatWithAgentErrors, CohereChatWithAgentResponses, CohereChatWithDefaultAgentData, CohereChatWithDefaultAgentErrors, CohereChatWithDefaultAgentResponses, CompleteOnboardingData, CompleteOnboardingErrors, CompleteOnboardingResponses, CreateAgentData, CreateAgentErrors, CreateAgentResponses, CreateApiKeyData, CreateApiKeyErrors, CreateApiKeyResponses, CreateChatConversationData, CreateChatConversationErrors, CreateChatConversationResponses, CreateChatOpsDmBindingData, CreateChatOpsDmBindingErrors, CreateChatOpsDmBindingResponses, CreateConnectorData, CreateConnectorErrors, CreateConnectorResponses, CreateInternalMcpCatalogItemData, CreateInternalMcpCatalogItemErrors, CreateInternalMcpCatalogItemResponses, CreateKnowledgeBaseData, CreateKnowledgeBaseErrors, CreateKnowledgeBaseResponses, CreateLimitData, CreateLimitErrors, CreateLimitResponses, CreateLlmProviderApiKeyData, CreateLlmProviderApiKeyErrors, CreateLlmProviderApiKeyResponses, CreateMcpServerInstallationRequestData, CreateMcpServerInstallationRequestErrors, CreateMcpServerInstallationRequestResponses, CreateOptimizationRuleData, CreateOptimizationRuleErrors, CreateOptimizationRuleResponses, CreateScheduleTriggerData, CreateScheduleTriggerErrors, CreateScheduleTriggerResponses, CreateScheduleTriggerRunConversationData, CreateScheduleTriggerRunConversationErrors, CreateScheduleTriggerRunConversationResponses, CreateTeamData, CreateTeamErrors, CreateTeamResponses, CreateToolInvocationPolicyData, CreateToolInvocationPolicyErrors, CreateToolInvocationPolicyResponses, CreateTrustedDataPolicyData, CreateTrustedDataPolicyErrors, CreateTrustedDataPolicyResponses, CreateVirtualApiKeyData, CreateVirtualApiKeyErrors, CreateVirtualApiKeyResponses, DeclineMcpServerInstallationRequestData, DeclineMcpServerInstallationRequestErrors, DeclineMcpServerInstallationRequestResponses, DeepseekChatCompletionsWithAgentData, DeepseekChatCompletionsWithAgentErrors, DeepseekChatCompletionsWithAgentResponses, DeepseekChatCompletionsWithDefaultAgentData, DeepseekChatCompletionsWithDefaultAgentErrors, DeepseekChatCompletionsWithDefaultAgentResponses, DeleteAgentData, DeleteAgentDelegationData, DeleteAgentDelegationErrors, DeleteAgentDelegationResponses, DeleteAgentErrors, DeleteAgentResponses, DeleteApiKeyData, DeleteApiKeyErrors, DeleteApiKeyResponses, DeleteChatConversationData, DeleteChatConversationErrors, DeleteChatConversationResponses, DeleteChatOpsBindingData, DeleteChatOpsBindingErrors, DeleteChatOpsBindingResponses, DeleteConnectorData, DeleteConnectorErrors, DeleteConnectorResponses, DeleteConversationEnabledToolsData, DeleteConversationEnabledToolsErrors, DeleteConversationEnabledToolsResponses, DeleteIncomingEmailSubscriptionData, DeleteIncomingEmailSubscriptionErrors, DeleteIncomingEmailSubscriptionResponses, DeleteInternalMcpCatalogItemByNameData, DeleteInternalMcpCatalogItemByNameErrors, DeleteInternalMcpCatalogItemByNameResponses, DeleteInternalMcpCatalogItemData, DeleteInternalMcpCatalogItemErrors, DeleteInternalMcpCatalogItemResponses, DeleteKnowledgeBaseData, DeleteKnowledgeBaseErrors, DeleteKnowledgeBaseResponses, DeleteLimitData, DeleteLimitErrors, DeleteLimitResponses, DeleteLlmProviderApiKeyData, DeleteLlmProviderApiKeyErrors, DeleteLlmProviderApiKeyResponses, DeleteMcpServerData, DeleteMcpServerErrors, DeleteMcpServerInstallationRequestData, DeleteMcpServerInstallationRequestErrors, DeleteMcpServerInstallationRequestResponses, DeleteMcpServerResponses, DeleteOptimizationRuleData, DeleteOptimizationRuleErrors, DeleteOptimizationRuleResponses, DeletePendingSignupMemberData, DeletePendingSignupMemberErrors, DeletePendingSignupMemberResponses, DeleteScheduleTriggerData, DeleteScheduleTriggerErrors, DeleteScheduleTriggerResponses, DeleteTeamData, DeleteTeamErrors, DeleteTeamResponses, DeleteToolData, DeleteToolErrors, DeleteToolInvocationPolicyData, DeleteToolInvocationPolicyErrors, DeleteToolInvocationPolicyResponses, DeleteToolResponses, DeleteTrustedDataPolicyData, DeleteTrustedDataPolicyErrors, DeleteTrustedDataPolicyResponses, DeleteVirtualApiKeyData, DeleteVirtualApiKeyErrors, DeleteVirtualApiKeyResponses, DisableScheduleTriggerData, DisableScheduleTriggerErrors, DisableScheduleTriggerResponses, DropEmbeddingConfigData, DropEmbeddingConfigErrors, DropEmbeddingConfigResponses, EnableScheduleTriggerData, EnableScheduleTriggerErrors, EnableScheduleTriggerResponses, ForceResyncConnectorData, ForceResyncConnectorErrors, ForceResyncConnectorResponses, ForkSharedConversationData, ForkSharedConversationErrors, ForkSharedConversationResponses, GenerateChatConversationTitleData, GenerateChatConversationTitleErrors, GenerateChatConversationTitleResponses, GetAgentData, GetAgentDelegationsData, GetAgentDelegationsErrors, GetAgentDelegationsResponses, GetAgentEmailAddressData, GetAgentEmailAddressErrors, GetAgentEmailAddressResponses, GetAgentErrors, GetAgentResponses, GetAgentsData, GetAgentsErrors, GetAgentsResponses, GetAgentStatisticsData, GetAgentStatisticsErrors, GetAgentStatisticsResponses, GetAgentToolsData, GetAgentToolsErrors, GetAgentToolsResponses, GetAllAgentsData, GetAllAgentsErrors, GetAllAgentsResponses, GetAllAgentToolsData, GetAllAgentToolsErrors, GetAllAgentToolsResponses, GetAllDelegationConnectionsData, GetAllDelegationConnectionsErrors, GetAllDelegationConnectionsResponses, GetAllVirtualApiKeysData, GetAllVirtualApiKeysErrors, GetAllVirtualApiKeysResponses, GetApiAuthBy__Data, GetApiAuthBy__Responses, GetApiAuthOauth2AuthorizeData, GetApiAuthOauth2AuthorizeResponses, GetApiKeyData, GetApiKeyErrors, GetApiKeyResponses, GetApiKeysData, GetApiKeysErrors, GetApiKeysResponses, GetAppearanceSettingsData, GetAppearanceSettingsErrors, GetAppearanceSettingsResponses, GetAvailableLlmProviderApiKeysData, GetAvailableLlmProviderApiKeysErrors, GetAvailableLlmProviderApiKeysResponses, GetChatAgentMcpToolsData, GetChatAgentMcpToolsErrors, GetChatAgentMcpToolsResponses, GetChatConversationData, GetChatConversationErrors, GetChatConversationResponses, GetChatConversationsData, GetChatConversationsErrors, GetChatConversationsResponses, GetChatOpsStatusData, GetChatOpsStatusErrors, GetChatOpsStatusResponses, GetConfigData, GetConfigResponses, GetConnectorData, GetConnectorErrors, GetConnectorKnowledgeBasesData, GetConnectorKnowledgeBasesErrors, GetConnectorKnowledgeBasesResponses, GetConnectorResponses, GetConnectorRunData, GetConnectorRunErrors, GetConnectorRunResponses, GetConnectorRunsData, GetConnectorRunsErrors, GetConnectorRunsResponses, GetConnectorsData, GetConnectorsErrors, GetConnectorsResponses, GetConversationEnabledToolsData, GetConversationEnabledToolsErrors, GetConversationEnabledToolsResponses, GetConversationShareData, GetConversationShareErrors, GetConversationShareResponses, GetCostSavingsStatisticsData, GetCostSavingsStatisticsErrors, GetCostSavingsStatisticsResponses, GetDefaultCredentialsStatusData, GetDefaultCredentialsStatusErrors, GetDefaultCredentialsStatusResponses, GetDefaultLlmProxyData, GetDefaultLlmProxyErrors, GetDefaultLlmProxyResponses, GetDefaultMcpGatewayData, GetDefaultMcpGatewayErrors, GetDefaultMcpGatewayResponses, GetDeploymentYamlPreviewData, GetDeploymentYamlPreviewErrors, GetDeploymentYamlPreviewResponses, GetHealthData, GetHealthResponses, GetIncomingEmailStatusData, GetIncomingEmailStatusErrors, GetIncomingEmailStatusResponses, GetInteractionData, GetInteractionErrors, GetInteractionResponses, GetInteractionsData, GetInteractionsErrors, GetInteractionSessionsData, GetInteractionSessionsErrors, GetInteractionSessionsResponses, GetInteractionsResponses, GetInternalMcpCatalogData, GetInternalMcpCatalogErrors, GetInternalMcpCatalogItemData, GetInternalMcpCatalogItemErrors, GetInternalMcpCatalogItemResponses, GetInternalMcpCatalogLabelKeysData, GetInternalMcpCatalogLabelKeysErrors, GetInternalMcpCatalogLabelKeysResponses, GetInternalMcpCatalogLabelValuesData, GetInternalMcpCatalogLabelValuesErrors, GetInternalMcpCatalogLabelValuesResponses, GetInternalMcpCatalogResponses, GetInternalMcpCatalogToolsData, GetInternalMcpCatalogToolsErrors, GetInternalMcpCatalogToolsResponses, GetK8sImagePullSecretsData, GetK8sImagePullSecretsErrors, GetK8sImagePullSecretsResponses, GetKnowledgeBaseData, GetKnowledgeBaseErrors, GetKnowledgeBaseHealthData, GetKnowledgeBaseHealthErrors, GetKnowledgeBaseHealthResponses, GetKnowledgeBaseResponses, GetKnowledgeBasesData, GetKnowledgeBasesErrors, GetKnowledgeBasesResponses, GetLabelKeysData, GetLabelKeysErrors, GetLabelKeysResponses, GetLabelValuesData, GetLabelValuesErrors, GetLabelValuesResponses, GetLimitData, GetLimitErrors, GetLimitResponses, GetLimitsData, GetLimitsErrors, GetLimitsResponses, GetLlmModelsData, GetLlmModelsErrors, GetLlmModelsResponses, GetLlmProviderApiKeyData, GetLlmProviderApiKeyErrors, GetLlmProviderApiKeyResponses, GetLlmProviderApiKeysData, GetLlmProviderApiKeysErrors, GetLlmProviderApiKeysResponses, GetMcpServerData, GetMcpServerErrors, GetMcpServerInstallationRequestData, GetMcpServerInstallationRequestErrors, GetMcpServerInstallationRequestResponses, GetMcpServerInstallationRequestsData, GetMcpServerInstallationRequestsErrors, GetMcpServerInstallationRequestsResponses, GetMcpServerInstallationStatusData, GetMcpServerInstallationStatusErrors, GetMcpServerInstallationStatusResponses, GetMcpServerResponses, GetMcpServersData, GetMcpServersErrors, GetMcpServersResponses, GetMcpServerToolsData, GetMcpServerToolsErrors, GetMcpServerToolsResponses, GetMcpToolCallData, GetMcpToolCallErrors, GetMcpToolCallResponses, GetMcpToolCallsData, GetMcpToolCallsErrors, GetMcpToolCallsResponses, GetMemberDefaultAgentData, GetMemberDefaultAgentErrors, GetMemberDefaultAgentResponses, GetMembersData, GetMembersErrors, GetMemberSignupStatusData, GetMemberSignupStatusErrors, GetMemberSignupStatusResponses, GetMembersResponses, GetModelStatisticsData, GetModelStatisticsErrors, GetModelStatisticsResponses, GetModelsWithApiKeysData, GetModelsWithApiKeysErrors, GetModelsWithApiKeysResponses, GetOAuthClientInfoData, GetOAuthClientInfoResponses, GetOnboardingStatusData, GetOnboardingStatusErrors, GetOnboardingStatusResponses, GetOperatorsData, GetOperatorsErrors, GetOperatorsResponses, GetOptimizationRulesData, GetOptimizationRulesErrors, GetOptimizationRulesResponses, GetOrganizationData, GetOrganizationErrors, GetOrganizationMemberData, GetOrganizationMemberErrors, GetOrganizationMemberResponses, GetOrganizationMembersData, GetOrganizationMembersErrors, GetOrganizationMembersResponses, GetOrganizationResponses, GetOverviewStatisticsData, GetOverviewStatisticsErrors, GetOverviewStatisticsResponses, GetPublicConfigData, GetPublicConfigResponses, GetReadyData, GetReadyErrors, GetReadyResponses, GetRoleData, GetRoleErrors, GetRoleResponses, GetRolesData, GetRolesErrors, GetRolesResponses, GetScheduleTriggerData, GetScheduleTriggerErrors, GetScheduleTriggerResponses, GetScheduleTriggerRunData, GetScheduleTriggerRunErrors, GetScheduleTriggerRunResponses, GetScheduleTriggerRunsData, GetScheduleTriggerRunsErrors, GetScheduleTriggerRunsResponses, GetScheduleTriggersData, GetScheduleTriggersErrors, GetScheduleTriggersResponses, GetSecretData, GetSecretErrors, GetSecretResponses, GetSecretsTypeData, GetSecretsTypeErrors, GetSecretsTypeResponses, GetSharedConversationData, GetSharedConversationErrors, GetSharedConversationResponses, GetTeamData, GetTeamErrors, GetTeamExternalGroupsData, GetTeamExternalGroupsErrors, GetTeamExternalGroupsResponses, GetTeamMembersData, GetTeamMembersErrors, GetTeamMembersResponses, GetTeamResponses, GetTeamsData, GetTeamsErrors, GetTeamsResponses, GetTeamStatisticsData, GetTeamStatisticsErrors, GetTeamStatisticsResponses, GetTokensData, GetTokensErrors, GetTokensResponses, GetTokenValueData, GetTokenValueErrors, GetTokenValueResponses, GetToolInvocationPoliciesData, GetToolInvocationPoliciesErrors, GetToolInvocationPoliciesResponses, GetToolInvocationPolicyData, GetToolInvocationPolicyErrors, GetToolInvocationPolicyResponses, GetToolsData, GetToolsErrors, GetToolsResponses, GetToolsWithAssignmentsData, GetToolsWithAssignmentsErrors, GetToolsWithAssignmentsResponses, GetTrustedDataPoliciesData, GetTrustedDataPoliciesErrors, GetTrustedDataPoliciesResponses, GetTrustedDataPolicyData, GetTrustedDataPolicyErrors, GetTrustedDataPolicyResponses, GetUniqueExternalAgentIdsData, GetUniqueExternalAgentIdsErrors, GetUniqueExternalAgentIdsResponses, GetUniqueUserIdsData, GetUniqueUserIdsErrors, GetUniqueUserIdsResponses, GetUserPermissionsData, GetUserPermissionsErrors, GetUserPermissionsResponses, GetUserStatisticsData, GetUserStatisticsErrors, GetUserStatisticsResponses, GetUserTokenData, GetUserTokenErrors, GetUserTokenResponses, GetUserTokenValueData, GetUserTokenValueErrors, GetUserTokenValueResponses, GetV1A2aByAgentIdWellKnownAgentJsonData, GetV1A2aByAgentIdWellKnownAgentJsonResponses, GetVirtualApiKeysData, GetVirtualApiKeysErrors, GetVirtualApiKeysResponses, GetVirtualKeyStatisticsData, GetVirtualKeyStatisticsErrors, GetVirtualKeyStatisticsResponses, GetWellKnownOauthAuthorizationServerData, GetWellKnownOauthAuthorizationServerResponses, GetWellKnownOauthProtectedResourceBy__Data, GetWellKnownOauthProtectedResourceBy__Responses, GroqChatCompletionsWithAgentData, GroqChatCompletionsWithAgentErrors, GroqChatCompletionsWithAgentResponses, GroqChatCompletionsWithDefaultAgentData, GroqChatCompletionsWithDefaultAgentErrors, GroqChatCompletionsWithDefaultAgentResponses, HandleOAuthCallbackData, HandleOAuthCallbackErrors, HandleOAuthCallbackResponses, InitiateOAuthData, InitiateOAuthErrors, InitiateOAuthResponses, InspectMcpServerData, InspectMcpServerErrors, InspectMcpServerResponses, InstallMcpServerData, InstallMcpServerErrors, InstallMcpServerResponses, ListChatOpsBindingsData, ListChatOpsBindingsErrors, ListChatOpsBindingsResponses, McpGatewayGetData, McpGatewayGetErrors, McpGatewayGetResponses, McpGatewayPostData, McpGatewayPostResponses, McpProxyPostData, McpProxyPostResponses, MinimaxChatCompletionsWithAgentData, MinimaxChatCompletionsWithAgentErrors, MinimaxChatCompletionsWithAgentResponses, MinimaxChatCompletionsWithDefaultAgentData, MinimaxChatCompletionsWithDefaultAgentErrors, MinimaxChatCompletionsWithDefaultAgentResponses, MistralChatCompletionsWithAgentData, MistralChatCompletionsWithAgentErrors, MistralChatCompletionsWithAgentResponses, MistralChatCompletionsWithDefaultAgentData, MistralChatCompletionsWithDefaultAgentErrors, MistralChatCompletionsWithDefaultAgentResponses, OllamaChatCompletionsWithAgentData, OllamaChatCompletionsWithAgentErrors, OllamaChatCompletionsWithAgentResponses, OllamaChatCompletionsWithDefaultAgentData, OllamaChatCompletionsWithDefaultAgentErrors, OllamaChatCompletionsWithDefaultAgentResponses, OpenAiChatCompletionsWithAgentData, OpenAiChatCompletionsWithAgentErrors, OpenAiChatCompletionsWithAgentResponses, OpenAiChatCompletionsWithDefaultAgentData, OpenAiChatCompletionsWithDefaultAgentErrors, OpenAiChatCompletionsWithDefaultAgentResponses, OpenrouterChatCompletionsWithAgentData, OpenrouterChatCompletionsWithAgentErrors, OpenrouterChatCompletionsWithAgentResponses, OpenrouterChatCompletionsWithDefaultAgentData, OpenrouterChatCompletionsWithDefaultAgentErrors, OpenrouterChatCompletionsWithDefaultAgentResponses, PerplexityChatCompletionsWithAgentData, PerplexityChatCompletionsWithAgentErrors, PerplexityChatCompletionsWithAgentResponses, PerplexityChatCompletionsWithDefaultAgentData, PerplexityChatCompletionsWithDefaultAgentErrors, PerplexityChatCompletionsWithDefaultAgentResponses, PostApiAuthBy__Data, PostApiAuthBy__Responses, PostApiAuthOauth2RegisterData, PostApiAuthOauth2RegisterResponses, PostApiAuthOauth2TokenData, PostApiAuthOauth2TokenResponses, PostApiAuthOrganizationRemoveMemberData, PostApiAuthOrganizationRemoveMemberResponses, PostApiAuthSignInSsoData, PostApiAuthSignInSsoResponses, PostApiWebhooksChatopsMsTeamsData, PostApiWebhooksChatopsMsTeamsErrors, PostApiWebhooksChatopsMsTeamsResponses, PostApiWebhooksChatopsSlackData, PostApiWebhooksChatopsSlackErrors, PostApiWebhooksChatopsSlackInteractiveData, PostApiWebhooksChatopsSlackInteractiveErrors, PostApiWebhooksChatopsSlackInteractiveResponses, PostApiWebhooksChatopsSlackResponses, PostApiWebhooksChatopsSlackSlashCommandData, PostApiWebhooksChatopsSlackSlashCommandErrors, PostApiWebhooksChatopsSlackSlashCommandResponses, PostApiWebhooksIncomingEmailData, PostApiWebhooksIncomingEmailErrors, PostApiWebhooksIncomingEmailResponses, PostV1A2aByAgentIdData, PostV1A2aByAgentIdResponses, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentData, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentErrors, PostV1GeminiByAgentIdV1BetaModelsByModelGenerateContentResponses, PostV1GeminiByAgentIdV1BetaModelsByModelStreamGenerateContentData, PostV1GeminiByAgentIdV1BetaModelsByModelStreamGenerateContentErrors, PostV1GeminiV1BetaModelsByModelGenerateContentData, PostV1GeminiV1BetaModelsByModelGenerateContentErrors, PostV1GeminiV1BetaModelsByModelGenerateContentResponses, PostV1GeminiV1BetaModelsByModelStreamGenerateContentData, PostV1GeminiV1BetaModelsByModelStreamGenerateContentErrors, ReauthenticateMcpServerData, ReauthenticateMcpServerErrors, ReauthenticateMcpServerResponses, RefreshChatOpsChannelDiscoveryData, RefreshChatOpsChannelDiscoveryErrors, RefreshChatOpsChannelDiscoveryResponses, ReinstallMcpServerData, ReinstallMcpServerErrors, ReinstallMcpServerResponses, RemoveTeamExternalGroupData, RemoveTeamExternalGroupErrors, RemoveTeamExternalGroupResponses, RemoveTeamMemberData, RemoveTeamMemberErrors, RemoveTeamMemberResponses, RenewIncomingEmailSubscriptionData, RenewIncomingEmailSubscriptionErrors, RenewIncomingEmailSubscriptionResponses, ResetDeploymentYamlData, ResetDeploymentYamlErrors, ResetDeploymentYamlResponses, RotateTokenData, RotateTokenErrors, RotateTokenResponses, RotateUserTokenData, RotateUserTokenErrors, RotateUserTokenResponses, RunScheduleTriggerNowData, RunScheduleTriggerNowErrors, RunScheduleTriggerNowResponses, SetupIncomingEmailWebhookData, SetupIncomingEmailWebhookErrors, SetupIncomingEmailWebhookResponses, ShareConversationData, ShareConversationErrors, ShareConversationResponses, StopChatStreamData, StopChatStreamErrors, StopChatStreamResponses, StreamChatData, StreamChatErrors, SubmitOAuthConsentData, SubmitOAuthConsentErrors, SubmitOAuthConsentResponses, SyncAgentDelegationsData, SyncAgentDelegationsErrors, SyncAgentDelegationsResponses, SyncConnectorData, SyncConnectorErrors, SyncConnectorResponses, SyncLlmModelsData, SyncLlmModelsErrors, SyncLlmModelsResponses, TestConnectorConnectionData, TestConnectorConnectionErrors, TestConnectorConnectionResponses, TestEmbeddingConnectionData, TestEmbeddingConnectionErrors, TestEmbeddingConnectionResponses, UnassignConnectorFromKnowledgeBaseData, UnassignConnectorFromKnowledgeBaseErrors, UnassignConnectorFromKnowledgeBaseResponses, UnassignToolFromAgentData, UnassignToolFromAgentErrors, UnassignToolFromAgentResponses, UnshareConversationData, UnshareConversationErrors, UnshareConversationResponses, UpdateAgentData, UpdateAgentErrors, UpdateAgentResponses, UpdateAgentSettingsData, UpdateAgentSettingsErrors, UpdateAgentSettingsResponses, UpdateAgentToolData, UpdateAgentToolErrors, UpdateAgentToolResponses, UpdateAppearanceSettingsData, UpdateAppearanceSettingsErrors, UpdateAppearanceSettingsResponses, UpdateChatConversationData, UpdateChatConversationErrors, UpdateChatConversationResponses, UpdateChatMessageData, UpdateChatMessageErrors, UpdateChatMessageResponses, UpdateChatOpsBindingData, UpdateChatOpsBindingErrors, UpdateChatOpsBindingResponses, UpdateChatOpsConfigInQuickstartData, UpdateChatOpsConfigInQuickstartErrors, UpdateChatOpsConfigInQuickstartResponses, UpdateConnectionSettingsData, UpdateConnectionSettingsErrors, UpdateConnectionSettingsResponses, UpdateConnectorData, UpdateConnectorErrors, UpdateConnectorResponses, UpdateConversationEnabledToolsData, UpdateConversationEnabledToolsErrors, UpdateConversationEnabledToolsResponses, UpdateInternalMcpCatalogItemData, UpdateInternalMcpCatalogItemErrors, UpdateInternalMcpCatalogItemResponses, UpdateKnowledgeBaseData, UpdateKnowledgeBaseErrors, UpdateKnowledgeBaseResponses, UpdateKnowledgeSettingsData, UpdateKnowledgeSettingsErrors, UpdateKnowledgeSettingsResponses, UpdateLimitData, UpdateLimitErrors, UpdateLimitResponses, UpdateLlmProviderApiKeyData, UpdateLlmProviderApiKeyErrors, UpdateLlmProviderApiKeyResponses, UpdateLlmSettingsData, UpdateLlmSettingsErrors, UpdateLlmSettingsResponses, UpdateMcpServerInstallationRequestData, UpdateMcpServerInstallationRequestErrors, UpdateMcpServerInstallationRequestResponses, UpdateMcpSettingsData, UpdateMcpSettingsErrors, UpdateMcpSettingsResponses, UpdateModelData, UpdateModelErrors, UpdateModelResponses, UpdateOptimizationRuleData, UpdateOptimizationRuleErrors, UpdateOptimizationRuleResponses, UpdateScheduleTriggerData, UpdateScheduleTriggerErrors, UpdateScheduleTriggerResponses, UpdateSecuritySettingsData, UpdateSecuritySettingsErrors, UpdateSecuritySettingsResponses, UpdateSlackChatOpsConfigData, UpdateSlackChatOpsConfigErrors, UpdateSlackChatOpsConfigResponses, UpdateTeamData, UpdateTeamErrors, UpdateTeamResponses, UpdateToolInvocationPolicyData, UpdateToolInvocationPolicyErrors, UpdateToolInvocationPolicyResponses, UpdateTrustedDataPolicyData, UpdateTrustedDataPolicyErrors, UpdateTrustedDataPolicyResponses, UpdateVirtualApiKeyData, UpdateVirtualApiKeyErrors, UpdateVirtualApiKeyResponses, ValidateDeploymentYamlData, ValidateDeploymentYamlErrors, ValidateDeploymentYamlResponses, VllmChatCompletionsWithAgentData, VllmChatCompletionsWithAgentErrors, VllmChatCompletionsWithAgentResponses, VllmChatCompletionsWithDefaultAgentData, VllmChatCompletionsWithDefaultAgentErrors, VllmChatCompletionsWithDefaultAgentResponses, XaiChatCompletionsWithAgentData, XaiChatCompletionsWithAgentErrors, XaiChatCompletionsWithAgentResponses, XaiChatCompletionsWithDefaultAgentData, XaiChatCompletionsWithDefaultAgentErrors, XaiChatCompletionsWithDefaultAgentResponses, ZhipuaiChatCompletionsWithAgentData, ZhipuaiChatCompletionsWithAgentErrors, ZhipuaiChatCompletionsWithAgentResponses, ZhipuaiChatCompletionsWithDefaultAgentData, ZhipuaiChatCompletionsWithDefaultAgentErrors, ZhipuaiChatCompletionsWithDefaultAgentResponses } from './types.gen';
 
 export type Options<TData extends TDataShape = TDataShape, ThrowOnError extends boolean = boolean> = Options2<TData, ThrowOnError> & {
     /**
@@ -3182,39 +3182,6 @@ export const updateOptimizationRule = <ThrowOnError extends boolean = false>(opt
 export const getRoles = <ThrowOnError extends boolean = false>(options?: Options<GetRolesData, ThrowOnError>) => (options?.client ?? client).get<GetRolesResponses, GetRolesErrors, ThrowOnError>({ url: '/api/roles', ...options });
 
 /**
- * Create a new custom role
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `ac:create`: Create new custom roles
- */
-export const createRole = <ThrowOnError extends boolean = false>(options: Options<CreateRoleData, ThrowOnError>) => (options.client ?? client).post<CreateRoleResponses, CreateRoleErrors, ThrowOnError>({
-    url: '/api/roles',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * Delete a custom role
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `ac:delete`: Delete custom roles
- */
-export const deleteRole = <ThrowOnError extends boolean = false>(options: Options<DeleteRoleData, ThrowOnError>) => (options.client ?? client).delete<DeleteRoleResponses, DeleteRoleErrors, ThrowOnError>({ url: '/api/roles/{roleId}', ...options });
-
-/**
  * Get a specific role by ID
  *
  * Authentication:
@@ -3226,26 +3193,6 @@ export const deleteRole = <ThrowOnError extends boolean = false>(options: Option
  * `ac:read`: View custom roles and their permissions
  */
 export const getRole = <ThrowOnError extends boolean = false>(options: Options<GetRoleData, ThrowOnError>) => (options.client ?? client).get<GetRoleResponses, GetRoleErrors, ThrowOnError>({ url: '/api/roles/{roleId}', ...options });
-
-/**
- * Update a custom role
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `ac:update`: Modify custom role permissions
- */
-export const updateRole = <ThrowOnError extends boolean = false>(options: Options<UpdateRoleData, ThrowOnError>) => (options.client ?? client).put<UpdateRoleResponses, UpdateRoleErrors, ThrowOnError>({
-    url: '/api/roles/{roleId}',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
 
 /**
  * Get organization details
@@ -4319,210 +4266,6 @@ export const zhipuaiChatCompletionsWithDefaultAgent = <ThrowOnError extends bool
  */
 export const zhipuaiChatCompletionsWithAgent = <ThrowOnError extends boolean = false>(options: Options<ZhipuaiChatCompletionsWithAgentData, ThrowOnError>) => (options.client ?? client).post<ZhipuaiChatCompletionsWithAgentResponses, ZhipuaiChatCompletionsWithAgentErrors, ThrowOnError>({
     url: '/v1/zhipuai/{agentId}/chat/completions',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * Get public identity provider list for login page (no secrets exposed)
- *
- * Authentication:
- *
- * Not required.
- *
- * Authorization:
- *
- * None (no additional RBAC permission required)
- */
-export const getPublicIdentityProviders = <ThrowOnError extends boolean = false>(options?: Options<GetPublicIdentityProvidersData, ThrowOnError>) => (options?.client ?? client).get<GetPublicIdentityProvidersResponses, GetPublicIdentityProvidersErrors, ThrowOnError>({ url: '/api/identity-providers/public', ...options });
-
-/**
- * Get all identity providers with full configuration (admin only)
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `identityProvider:read`: View identity provider configurations (SSO)
- */
-export const getIdentityProviders = <ThrowOnError extends boolean = false>(options?: Options<GetIdentityProvidersData, ThrowOnError>) => (options?.client ?? client).get<GetIdentityProvidersResponses, GetIdentityProvidersErrors, ThrowOnError>({ url: '/api/identity-providers', ...options });
-
-/**
- * Create a new identity provider
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `identityProvider:create`: Set up new identity providers
- */
-export const createIdentityProvider = <ThrowOnError extends boolean = false>(options: Options<CreateIdentityProviderData, ThrowOnError>) => (options.client ?? client).post<CreateIdentityProviderResponses, CreateIdentityProviderErrors, ThrowOnError>({
-    url: '/api/identity-providers',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * Get the IdP logout URL for the current user's identity provider
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * None (no additional RBAC permission required)
- */
-export const getIdentityProviderIdpLogoutUrl = <ThrowOnError extends boolean = false>(options?: Options<GetIdentityProviderIdpLogoutUrlData, ThrowOnError>) => (options?.client ?? client).get<GetIdentityProviderIdpLogoutUrlResponses, GetIdentityProviderIdpLogoutUrlErrors, ThrowOnError>({ url: '/api/identity-providers/idp-logout-url', ...options });
-
-/**
- * Delete identity provider
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `identityProvider:delete`: Remove identity providers
- */
-export const deleteIdentityProvider = <ThrowOnError extends boolean = false>(options: Options<DeleteIdentityProviderData, ThrowOnError>) => (options.client ?? client).delete<DeleteIdentityProviderResponses, DeleteIdentityProviderErrors, ThrowOnError>({ url: '/api/identity-providers/{id}', ...options });
-
-/**
- * Get identity provider by ID
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `identityProvider:read`: View identity provider configurations (SSO)
- */
-export const getIdentityProvider = <ThrowOnError extends boolean = false>(options: Options<GetIdentityProviderData, ThrowOnError>) => (options.client ?? client).get<GetIdentityProviderResponses, GetIdentityProviderErrors, ThrowOnError>({ url: '/api/identity-providers/{id}', ...options });
-
-/**
- * Update identity provider
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `identityProvider:update`: Modify identity provider settings
- */
-export const updateIdentityProvider = <ThrowOnError extends boolean = false>(options: Options<UpdateIdentityProviderData, ThrowOnError>) => (options.client ?? client).put<UpdateIdentityProviderResponses, UpdateIdentityProviderErrors, ThrowOnError>({
-    url: '/api/identity-providers/{id}',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * Delete a team's Vault folder mapping
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:update`: Modify team settings
- */
-export const deleteTeamVaultFolder = <ThrowOnError extends boolean = false>(options: Options<DeleteTeamVaultFolderData, ThrowOnError>) => (options.client ?? client).delete<DeleteTeamVaultFolderResponses, DeleteTeamVaultFolderErrors, ThrowOnError>({ url: '/api/teams/{teamId}/vault-folder', ...options });
-
-/**
- * Get a team's Vault folder configuration
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:read`: View teams and their members
- */
-export const getTeamVaultFolder = <ThrowOnError extends boolean = false>(options: Options<GetTeamVaultFolderData, ThrowOnError>) => (options.client ?? client).get<GetTeamVaultFolderResponses, GetTeamVaultFolderErrors, ThrowOnError>({ url: '/api/teams/{teamId}/vault-folder', ...options });
-
-/**
- * Set or update a team's Vault folder path
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:update`: Modify team settings
- */
-export const setTeamVaultFolder = <ThrowOnError extends boolean = false>(options: Options<SetTeamVaultFolderData, ThrowOnError>) => (options.client ?? client).post<SetTeamVaultFolderResponses, SetTeamVaultFolderErrors, ThrowOnError>({
-    url: '/api/teams/{teamId}/vault-folder',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * Check connectivity to a team's Vault folder
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:update`: Modify team settings
- */
-export const checkTeamVaultFolderConnectivity = <ThrowOnError extends boolean = false>(options: Options<CheckTeamVaultFolderConnectivityData, ThrowOnError>) => (options.client ?? client).post<CheckTeamVaultFolderConnectivityResponses, CheckTeamVaultFolderConnectivityErrors, ThrowOnError>({
-    url: '/api/teams/{teamId}/vault-folder/check-connectivity',
-    ...options,
-    headers: {
-        'Content-Type': 'application/json',
-        ...options.headers
-    }
-});
-
-/**
- * List secrets available in a team's Vault folder
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:read`: View teams and their members
- */
-export const listTeamVaultFolderSecrets = <ThrowOnError extends boolean = false>(options: Options<ListTeamVaultFolderSecretsData, ThrowOnError>) => (options.client ?? client).get<ListTeamVaultFolderSecretsResponses, ListTeamVaultFolderSecretsErrors, ThrowOnError>({ url: '/api/teams/{teamId}/vault-folder/secrets', ...options });
-
-/**
- * Get the keys of a specific secret in a team's Vault folder
- *
- * Authentication:
- *
- * Required. Use an authenticated browser session or send your Archestra API key in the `Authorization` header.
- *
- * Authorization:
- *
- * `team:read`: View teams and their members
- */
-export const getTeamVaultSecretKeys = <ThrowOnError extends boolean = false>(options: Options<GetTeamVaultSecretKeysData, ThrowOnError>) => (options.client ?? client).post<GetTeamVaultSecretKeysResponses, GetTeamVaultSecretKeysErrors, ThrowOnError>({
-    url: '/api/teams/{teamId}/vault-folder/secrets/keys',
     ...options,
     headers: {
         'Content-Type': 'application/json',
